% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_db.R
\name{clean_db}
\alias{clean_db}
\title{Clean DB}
\usage{
clean_db(credentials_file = "credentials/credentials.yaml")
}
\arguments{
\item{credentials_file}{Path to a credentials.yaml}
}
\value{
Data will be modified in the database
}
\description{
Clean the database
}
