% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_borders.R
\name{create_borders}
\alias{create_borders}
\title{Create borders}
\usage{
create_borders(df)
}
\arguments{
\item{df}{A dataframe with variables named \code{x}, \code{y}, and \code{cluster}}
}
\value{
A SpatialPolygonsDataFrame
}
\description{
Generate borders around clusters of points
}
