% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_buffers.R
\name{create_buffers}
\alias{create_buffers}
\title{Create buffers}
\usage{
create_buffers(shp, meters = 1000, crs = "+init=epsg:3347")
}
\arguments{
\item{shp}{The spatialPolygonsDataFrame (projected in lat/long)}

\item{meters}{The size of the buffer in meters}

\item{crs}{}
}
\value{
A SpatialPolygonsDataFrame
}
\description{
Generate buffers of size \code{meters} around points or polygons
}
