% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_clean_db.R
\name{create_clean_db}
\alias{create_clean_db}
\title{Create clean DB}
\usage{
create_clean_db(
  credentials_file = "credentials/credentials.yaml",
  drop_all = FALSE
)
}
\arguments{
\item{credentials_file}{Path to a credentials.yaml}

\item{drop_all}{Whether to, instead of creating clean tables, drop them}
}
\value{
Data will be created in the database
}
\description{
Create a clean database as a copy of the ODK "raw" database
}
