% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_clusters.R
\name{create_clusters}
\alias{create_clusters}
\title{Optimization of cluster formation}
\usage{
create_clusters(cluster_size = 10, plot_map = FALSE, sleep = 0,
  locations, shp = NULL, start = "close", rest = "close",
  messaging = FALSE, save = NULL)
}
\arguments{
\item{cluster_size}{The size of each cluster}

\item{plot_map}{Whether to plot a map after each cluster formation (slows down operations significantly)}

\item{sleep}{How many seconds to sleep between plots}

\item{locations}{a SpatialPointsDataFrame containing 3 columns: id, lat, and lon}

\item{shp}{The shapefile to plot (only if plot_map is TRUE)}

\item{start}{The start strategy}

\item{rest}{The rest strategy}

\item{messaging}{Whether to message or not, boolean}

\item{save}{Where to save pngs (if null, no saving)}

\item{locations}{locations}
}
\value{
A dataframe of size \code{times} times the number of points in \code{locations}, with columns indicating each \code{simulation_number}
}
\description{
Use the brute force method to assign households to clusters in a way that minimizes inter-cluster distance.
}
