% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_qr.R
\name{create_qr}
\alias{create_qr}
\title{Create QR}
\usage{
create_qr(id = "01234567")
}
\arguments{
\item{id}{The id number for which the qr code is being created}
}
\value{
A QR code
}
\description{
Generate a QR code for an ID number
}
