% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decrypt_private_data.R
\name{decrypt_private_data}
\alias{decrypt_private_data}
\title{Decrypt data provided using the provided private key}
\usage{
decrypt_private_data(data, keyfile)
}
\arguments{
\item{data}{The private data to be decrypted}

\item{keyfile}{The filename where the key to be used is stored}
}
\value{
A plain text string
}
\description{
Decrypt data provided using the provided private key
}
