% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encrypt_private_data.R
\name{encrypt_private_data}
\alias{encrypt_private_data}
\title{Encrypt data provided using the provided public key}
\usage{
encrypt_private_data(data, keyfile)
}
\arguments{
\item{data}{The private data to be encrypted}

\item{keyfile}{The filename (plus full file path) where the key to be used is stored}
}
\value{
A 256 char encoded string
}
\description{
Encrypt data provided using the provided public key
}
