% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enroll_worker.R
\name{enroll_worker}
\alias{enroll_worker}
\title{Enroll worker}
\usage{
enroll_worker(name_first = "Jane", name_last = "Doe", location = "Tanzania")
}
\arguments{
\item{name_first}{First name of the worker}

\item{name_last}{Last name of the worker}

\item{location}{Country of the worker (either Tanzania or Mozambique)}
}
\value{
A QR code
}
\description{
Enroll a new worker by (a) inputting information about that worker and (b) receiving an ID number and a list of households. By running this function, you will both create one new row in the "workers" table of the "ids" database, and 1000 new rows in the "households" table of the "ids" database
}
