% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_enumerations.R
\name{format_enumerations}
\alias{format_enumerations}
\title{Format enumerations ODK data for the bohemia database}
\usage{
format_enumerations(data)
}
\arguments{
\item{data}{The list of dataframes (non_repeats and repeats) generated by odk_get_data for enumerations}
}
\value{
A list of dataframes formatted and named in accordance with the bohemia database
}
\description{
Take the enumerations data (as generated by "odk_get_data") and format it so it's ready for popping into the enumerations table of the bohemia database
}
