% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_minicensus.R
\name{format_minicensus}
\alias{format_minicensus}
\title{Format minicensus ODK data for the bohemia database}
\usage{
format_minicensus(data)
}
\arguments{
\item{data}{The list of dataframes (non_repeats and repeats) generated by odk_get_data}
}
\value{
A list of dataframes formatted and named in accordance with the bohemia database
}
\description{
Take the minicensus data (as generated by "odk_get_data") and format it so it's ready for popping into the minicensus database
}
