% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_refusals.R
\name{format_refusals}
\alias{format_refusals}
\title{Format refusals ODK data for the bohemia database}
\usage{
format_refusals(data)
}
\arguments{
\item{data}{The list of dataframes (non_repeats and repeats) generated by odk_get_data for refusals}
}
\value{
A list of dataframes formatted and named in accordance with the bohemia database
}
\description{
Take the refusals data (as generated by "odk_get_data") and format it so it's ready for popping into the refusals table of the bohemia database
}
