% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_va.R
\name{format_va}
\alias{format_va}
\title{Format va ODK data for the bohemia database}
\usage{
format_va(data)
}
\arguments{
\item{data}{The list of dataframes (non_repeats and repeats) generated by odk_get_data for va}
}
\value{
A list of dataframes formatted and named in accordance with the bohemia database
}
\description{
Take the va data (as generated by "odk_get_data") and format it so it's ready for popping into the va table of the bohemia database
}
