% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_va_refusals.R
\name{format_va_refusals}
\alias{format_va_refusals}
\title{Format VA Refusals}
\usage{
format_va_refusals(data, keyfile)
}
\arguments{
\item{data}{The list of dataframes (non_repeats and repeats) generated by odk_get_data}

\item{keyfile}{The public key file to be used for encryption}
}
\value{
A list of dataframes formatted and named in accordance with the bohemia database
}
\description{
Take the VA Refusals data (as generated by "odk_get_data") and format it so it's ready for popping into the database
}
