% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keygen_rsa.R
\name{generate_keys}
\alias{generate_keys}
\title{Generate public private key pairs}
\usage{
generate_keys(directory = getwd())
}
\arguments{
\item{directory}{Which file to put the keys into}
}
\value{
A pair of pem key files
}
\description{
Generate public private key pairs
}
