% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_location_codes.R
\name{generate_location_codes}
\alias{generate_location_codes}
\title{Generate location codes}
\usage{
generate_location_codes(locations)
}
\arguments{
\item{locations}{A dataframe in the format of the locations hierarchy (ie, the bohemia::locations object, with the following 6 columns: Country, region, District, Ward, Village, Hamlet)}
}
\value{
A dataframe in the input format, but with an additional "code" column, and columns indicating the degrees of change required
}
\description{
Generate unique 3 letter location given a locations hierarchy
}
