% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_briefcase.R
\name{get_data_briefcase}
\alias{get_data_briefcase}
\title{Get data briefcase}
\usage{
get_data_briefcase(
  url,
  id,
  user,
  password,
  briefcase_dir,
  jar_file_briefcase,
  briefcase_storage_dir,
  dry_run = FALSE
)
}
\arguments{
\item{url}{Country}

\item{id}{Region}

\item{user}{District}

\item{password}{Ward}

\item{briefcase_dir}{Village}

\item{jar_file_briefcase}{Hamlet}

\item{briefcase_storage_dir}{Allow the all option}

\item{dry_run}{Whether to just print the java cli code (rather than actually run it)}
}
\value{
Writes objects
}
\description{
Get the unique 3 letter location given a location
}
