% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_location_code.R
\name{get_location_code}
\alias{get_location_code}
\title{Get location code}
\usage{
get_location_code(
  country = NULL,
  region = NULL,
  district = NULL,
  ward = NULL,
  village = NULL,
  hamlet = NULL,
  allow_all = FALSE
)
}
\arguments{
\item{country}{Country}

\item{region}{Region}

\item{district}{District}

\item{ward}{Ward}

\item{village}{Village}

\item{hamlet}{Hamlet}

\item{allow_all}{Allow the all option}
}
\value{
A character vector of length 1 or a NULL object
}
\description{
Get the unique 3 letter location given a location
}
