% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_positions_from_unique_id.R
\name{get_positions_from_unique_id}
\alias{get_positions_from_unique_id}
\title{Get positions from unique id}
\usage{
get_positions_from_unique_id(url, user, pass, unique_id)
}
\arguments{
\item{url}{The url of your traccar api}

\item{user}{The username for your traccar api}

\item{pass}{The password for your traccar api}

\item{unique_id}{The unique id for a regestered HCW on the traccar api}
}
\value{
A dataframe
}
\description{
Retrieve positions using a unique id
}
