% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_to_odk.R
\name{google_to_odk}
\alias{google_to_odk}
\title{Google to ODK}
\usage{
google_to_odk(
 
    url = "https://docs.google.com/spreadsheets/d/16_drw-35haLaBlB6tn92mr6zbIuYorAUDyieGONyGTM/edit#gid=141178862",
  briefcase = "~/Documents/briefcase/ODK-Briefcase-v1.16.1.jar",
  form_id = NULL,
  form_title = NULL,
  storage_directory = "~/Documents/briefcase",
  aggregate_url = "https://bohemia.systems",
  odk_username = "data",
  odk_password = "data",
  download_as = "temp.xlsx",
  convert_to = "temp.xml"
)
}
\arguments{
\item{url}{The url of the google sheet}

\item{briefcase}{Location on the local machine of the briefcase jar file}

\item{download_as}{The file location to be downloaded to}

\item{convert_to}{The file location of the converted xml}

\item{aggregate}{Location of the ODK Aggregate server}
}
\value{
Nothing
}
\description{
Fetch an xlsform from google docs, convert to xml, and push to an ODK Aggregate server. The google sheet must have been "published to the web" via the file menu in the upper-left. Note, before running this script, one should have installed ODK Briefcase per the instructions at https://github.com/databrew/bohemia/blob/master/guides/guide_briefcase.md. Note, this is unstable, and proper functioning still requires manually deleting form definitions from the Aggregate server.
}
