% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hh_qr_code_print.R
\name{hh_qr_code_print}
\alias{hh_qr_code_print}
\title{Household id to qr code print}
\usage{
hh_qr_code_print(
  base_img_path,
  hh_id,
  demo_sig,
  pi_sig,
  country,
  save_file = NULL,
  ...
)
}
\arguments{
\item{base_img_path}{The path to the image that will be overlayed with a qr code}

\item{hh_id}{The household ID}

\item{pi_sig}{The path to the PI signature file}

\item{save_file}{Where to save the file}

\item{demog_sig}{The path to the demography signature file}
}
\value{
overlayed image
}
\description{
Overlay qr code on image
}
