% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hh_qr_code_simple_print.R
\name{hh_qr_code_print_simple}
\alias{hh_qr_code_print_simple}
\title{Household id to qr code print (simple)}
\usage{
hh_qr_code_print_simple(hh_id = "ABC-123", save_file = NULL, ...)
}
\arguments{
\item{hh_id}{The household ID}

\item{save_file}{Where to save the file}
}
\value{
overlayed image
}
\description{
Overlay qr code on image
}
