% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_anomalies_and_errors.R
\name{identify_anomalies_and_errors}
\alias{identify_anomalies_and_errors}
\title{Identify anomalies and errors}
\usage{
identify_anomalies_and_errors(data_list)
}
\arguments{
\item{data_list}{A named list of dataframes. It should have the structure of named dataframes sharing names with the names of the tables in the database (excluding the "clean_" prefix)}
}
\value{
Data will be modified in the database
}
\description{
Given a clean Bohemia database, identify anomalies and errors
}
