% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keygen_rsa.R
\docType{data}
\name{key}
\alias{key}
\title{Generate public private key pairs}
\format{
An object of class \code{public.key} (inherits from \code{private.key}) of length 1.
}
\usage{
key
}
\value{
A pair of pem key files
}
\description{
Generate public private key pairs
}
\keyword{datasets}
