% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_generation_file_index.R
\name{list_generation_file_index}
\alias{list_generation_file_index}
\title{Visit control sheet file index list}
\usage{
list_generation_file_index(location = NULL, output_file = NULL)
}
\arguments{
\item{location}{A three letter location code (or character vector of multiple codes). If NULL, all locations to be used}

\item{output_file}{The name of the spreadsheet to be written (csv file); if null, will return the table in memory}
}
\value{
A table will be created
}
\description{
Render a file index list
}
