% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_generation_va.R
\name{list_generation_va}
\alias{list_generation_va}
\title{VA list generation}
\usage{
list_generation_va(
  keyfile = NULL,
  keyfile_public = NULL,
  location = NULL,
  output_file = NULL
)
}
\arguments{
\item{keyfile}{The path to the private key}

\item{keyfile_public}{The path to the public key}

\item{location}{A three letter location code (or character vector of multiple codes). If NULL, all locations to be used}

\item{output_file}{The name of the spreadsheet to be written (csv file); if null, will return the table in memory}
}
\value{
A table will be created
}
\description{
Render a VA list
}
