% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_odk_data.R
\name{load_odk_data}
\alias{load_odk_data}
\title{Load ODK data}
\usage{
load_odk_data(
  the_country = "Mozambique",
  credentials_path = "credentials/credentials.yaml",
  users_path = "credentials/users.yaml",
  local = FALSE,
  efficient = TRUE,
  use_cached = FALSE,
  con = NULL
)
}
\arguments{
\item{the_country}{The country to be loaded}

\item{credentials_path}{The path to the credentials.yaml file}

\item{users_path}{The path to the users.yaml file}

\item{local}{Whether to use the local database}

\item{efficient}{Whether to only load select columns}

\item{use_cached}{Whether to use a cached file}

\item{con}{A connection object}
}
\description{
Load odk data
}
