% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odk_create_location_choices.R
\name{odk_create_location_choices}
\alias{odk_create_location_choices}
\title{Create location choices}
\usage{
odk_create_location_choices(country = NULL, add_other = FALSE,
  add_ids = TRUE, other_word = "Other")
}
\arguments{
\item{country}{Which country to use. If \code{NULL} (the default), both will be used.}
}
\value{
A list of two tables named "survey" and "choices"
}
\description{
Generate cascading options for location choices appropriate for the correct location hierarchy in ODK
}
