% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odk_get_secondary_id.R
\name{odk_get_secondary_id}
\alias{odk_get_secondary_id}
\title{ODK get secondary ID}
\usage{
odk_get_secondary_id(url = "https://bohemia.systems", id = "recon")
}
\arguments{
\item{url}{The URL of the ODK Aggregate server, default being https://bohemia.systems, without a trailing dash}

\item{id}{The primary id fo the form}
}
\value{
A character vector of length 1
}
\description{
Get the secondary ID of an ODK form list from an aggregate server. This is useful in the rare case that the xls used to generate the xml was not named identically to the \code{formId}. The secondary ID is used for inserting into a \code{downloadSubmission} http request (see \code{odk_get_submission})
}
