% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odk_get_submission.R
\name{odk_get_submission}
\alias{odk_get_submission}
\title{ODK get submission}
\usage{
odk_get_submission(
  url = "https://bohemia.systems",
  id = "recon",
  id2 = NULL,
  uuid = NULL,
  user = NULL,
  password = NULL
)
}
\arguments{
\item{url}{The URL of the ODK Aggregate server, default being https://bohemia.systems, without a trailing dash}

\item{id}{The primary id of the form}

\item{id2}{The secondary id of the form}

\item{uuid}{The uuid (including the "uuid:" prefix) of the particular submission to be retrieved}

\item{user}{The ODK Aggregate username}

\item{password}{The ODK Aggregate password}
}
\value{
An xml of the submission's content
}
\description{
Get the data from a particular submission
}
