% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odk_list_forms.R
\name{odk_list_forms}
\alias{odk_list_forms}
\title{List the forms on an ODK server}
\usage{
odk_list_forms(url = "https://bohemia.systems", user = NULL, password = NULL)
}
\arguments{
\item{url}{The URL of the ODK Aggregate server, default being https://bohemia.systems, without a trailing dash}

\item{user}{The ODK Aggregate username}

\item{password}{The ODK Aggregate password}
}
\value{
A tibble with columns name, id, and url
}
\description{
List the forms on an ODK server
}
