% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odk_list_submissions.R
\name{odk_list_submissions}
\alias{odk_list_submissions}
\title{ODK list submissions}
\usage{
odk_list_submissions(
  url = "https://bohemia.systems",
  id = "recon",
  user = NULL,
  password = NULL,
  pre_auth = FALSE,
  chunk_size = 1000
)
}
\arguments{
\item{url}{The URL of the ODK Aggregate server, default being https://bohemia.systems, without a trailing dash}

\item{id}{The primary id fo the form}

\item{user}{The ODK Aggregate username}

\item{password}{The ODK Aggregate password}

\item{chunk_size}{The chunk size for pagination}

\item{pre-auth}{Pre-authenticate (needed for Manhica server)}
}
\value{
A character vector of length 1 with the uuids (including the prefix "uuid:") of the submitted forms
}
\description{
Get a list of the submissions of a particular form ID to a particular server
}
