% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odk_make_wide.R
\name{odk_make_wide}
\alias{odk_make_wide}
\title{ODK make wide}
\usage{
odk_make_wide(long_list)
}
\arguments{
\item{long_list}{A list generated by \code{odk_parse_submission}}
}
\value{
A list of wide dataframes of identical length to the input. The \code{repeats} element of the list will be of length 1 or greater (depending on the number of repeats in the form)
}
\description{
Take a list of long datasets (as generated by \code{odk_parse_submission}) and convert to "wide" format
}
