% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odk_parse_submission.R
\name{odk_parse_submission}
\alias{odk_parse_submission}
\title{ODK parse submission}
\usage{
odk_parse_submission(xml)
}
\value{
A character vector of length 1 with the uuids (including the prefix "uuid:") of the submitted forms
}
\description{
Convert the xml data from an ODK form submission to a dataframe
}
\examples{
# Get a list of forms
fl <- odk_list_forms()
# # Keep only the one named "Recon"
recon <- fl \%>\% filter(name == 'Recon')
# Capture its secondary idea (slow)
id2 <- odk_get_secondary_id(id = recon$id)
# Get a list of the submissions for the form in question
submissions <- odk_list_submissions(url = 'https://bohemia.systems', id = 'recon', user = 'data', password = 'data')
# Capture an individual submission
submission <- odk_get_submission(url = 'https://bohemia.systems', id = recon$id, id2 = id2, uuid = submissions[1], user = 'data', password = 'data')
#' # Parse the submission into R format
#' odk_parse_submission(xml = submission)
}
