% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odk_to_codebook.R
\name{odk_to_codebook}
\alias{odk_to_codebook}
\title{ODK to codebook}
\usage{
odk_to_codebook(excel_file, docx_out, csv_out)
}
\arguments{
\item{excel_file}{The xls to read}

\item{docx_out}{The doc to write}

\item{csv_out}{The csv to write}
}
\value{
Writes files
}
\description{
Convert a xls to a codebook
}
