% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openhds_create_location_hierarchy.R
\name{openhds_create_location_hierarchy}
\alias{openhds_create_location_hierarchy}
\title{Create location hierarchy}
\usage{
openhds_create_location_hierarchy(output_file = "../scripts/locations")
}
\arguments{
\item{output_file}{Where to write the file (.sql and .csv will be appended)}
}
\value{
SQL code (in a .sql file) and tabular data (in a .csv file). The sql code is meant to be run on the OpenHDS database, and the .csv file is meant to be copy-pasted into the census choices
}
\description{
Generate cascading options for location hierarchy for insertion into openhds database and the ODK "choices" sheet. This assumes that location levels have been configured in the openhds web app under "Configuration"->"Location Levels", and have been set to the following: Country, District, Ward, Village, Hamlet (1-5).
}
