% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_traccar_data.R
\name{post_traccar_data}
\alias{post_traccar_data}
\title{Post traccar data}
\usage{
post_traccar_data(user, pass, name, unique_id, url)
}
\arguments{
\item{user}{The username for your traccar database}

\item{pass}{The password for your traccar database}

\item{name}{The name of the individual you are adding to your traccar server}

\item{unique_id}{The id of the individual you are adding to your traccar server}

\item{url}{The url of your traccar server}
}
\value{
A dataframe
}
\description{
Upload data to your traccar server
}
