% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_worker_qrs.R
\name{print_worker_qrs}
\alias{print_worker_qrs}
\title{Print worker QRs}
\usage{
print_worker_qrs(
  wid = 1,
  restrict = NULL,
  output_dir = NULL,
  output_file = "qrs.pdf"
)
}
\arguments{
\item{wid}{The 3 digit id of the worker}

\item{restrict}{Restrict to only some ids. If
\code{NULL} (the default), all 1000 ids assigned to the worker will be used. Otherwise, this should be a vector of either (a) the three character subids to be kept or (b) the 7 character hhids to be kept (3 digits, dash, 3 digits)}

\item{output_dir}{The directory to which the file should be written. If
\code{NULL} (the default), the current working directory will be used.}

\item{output_file}{The name of the file to be written.}
}
\value{
A pdf will be written
}
\description{
Render a document of pds for a worker
}
