% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_db.R
\name{read_db}
\alias{read_db}
\title{Read DB}
\usage{
read_db(crednetials_file = "credentials/credentials.yaml")
}
\arguments{
\item{credentials_file}{Path to a credentials.yaml}
}
\value{
A list of dataframes in memory
}
\description{
Read data from the Bohemia AWS RDS database
}
