% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_qr_pdf.R
\name{render_qr_pdf}
\alias{render_qr_pdf}
\title{Render QR pdf}
\usage{
render_qr_pdf(ids = NULL, output_dir = NULL,
  output_file = "paper.pdf")
}
\arguments{
\item{ids}{A character vector of ID numbers. If \code{NULL} (the
default), all IDs will be used}

\item{output_dir}{The directory to which the file should be written. If
\code{NULL} (the default), the current working directory will be used.}

\item{output_file}{The name of the file to be written.}
}
\value{
An html will be written
}
\description{
Render a pdf of QR codes
}
