% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_operations_app.R
\name{run_operations_app}
\alias{run_operations_app}
\title{Run operations app}
\usage{
run_operations_app()
}
\arguments{
\item{x}{The vector to be modified}

\item{n}{The number of characters that the resulting vector should have}
}
\value{
A character vector of identical length to \code{x} in which all elements have n characters (or more, for those which already had more prior to processing)
}
\description{
Run the Bohemia operations app
}
