% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up_local_db.R
\name{set_up_local_db}
\alias{set_up_local_db}
\title{Set up local DB}
\usage{
set_up_local_db(
  credentials_path = "credentials/credentials.yaml",
  file = "~/Desktop/bohemia.sql"
)
}
\arguments{
\item{credentials_path}{The path to the credentials.yaml file}
}
\description{
Set up local database (ie, copy remote)
}
