% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_workers_traccar.R
\name{sync_workers_traccar}
\alias{sync_workers_traccar}
\title{Sync workers traccar}
\usage{
sync_workers_traccar(
  credentials = NULL,
  delete_from_traccar = FALSE,
  traccar_url,
  traccar_user,
  traccar_pass
)
}
\arguments{
\item{credentials}{A list with named elements. If \code{NULL} these arguments can be specified individually}

\item{delete_from_traccar}{Whether to delete from Traccar those workers who don't show up in the shiny database. Be careful with this!}

\item{traccar_url}{The URL where the Traccar server application is being run}

\item{traccar_user}{The administrator's username for the Traccar server application}

\item{traccar_pass}{The administrator's password for the Traccar server application}
}
\value{
The function will send updates from the Shiny-generated MySQL database to the Traccar server via an HTTP post request
}
\description{
Syncronize between (a) the workers who have registered via google sheets and (b) the workers whose information has been registered on the Traccar server app
}
\examples{
\dontrun{
# Read in credentials
credentials <- yaml::yaml.load_file('../../../credentials/credentials.yaml')
sync_workers(credentials = credentials)
}
}
