% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_clusters.R
\name{try_clusters}
\alias{try_clusters}
\title{Try creating clusters}
\usage{
try_clusters(
  the_country = "Tanzania",
  include_clinical = TRUE,
  interpolate_animals = TRUE,
  interpolate_humans = TRUE,
  humans_per_household = 5,
  p_children = 30,
  minimum_households = 0,
  minimum_children = 35,
  minimum_humans = 0,
  minimum_animals = 35,
  minimum_cattle = 0,
  minimum_pigs = 0,
  minimum_goats = 0,
  df = NULL
)
}
\arguments{
\item{the_country}{Either Mozambique or Tanzania}

\item{include_clinical}{Whether to include hamlets which are part of another ongoing clinical trial}

\item{interpolate_animals}{Whether to guess the number of animals if they are missing}

\item{interpolate_humans}{Whether to guess the number of humans if they are missing}

\item{humans_per_household}{The number of humans per household}

\item{p_children}{The percentage of people which are considered to be children}

\item{minimum_children}{The minimum number of children required for a cluster}

\item{minimum_humans}{The minimum number of humans required for a cluster}

\item{minimum_animals}{The minimum number of animals required for a cluster}

\item{minimum_cattle}{The minimum number of cattle required for a cluster}

\item{minimum_pigs}{The minimum number of pigs required for a cluster}

\item{minimum_goats}{The minimum number of goats required for a cluster}

\item{df}{A dataframe with combined animal and household info, as generated in global.R of operations app}
}
\value{
A list
}
\description{
Try creating clusters from the recon / animal annex data
}
