% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_enumerations.R
\name{update_enumerations}
\alias{update_enumerations}
\title{Update enumerations data in the bohemia database}
\usage{
update_enumerations(
  formatted_data,
  con = NULL,
  host = NULL,
  user = NULL,
  pass = NULL,
  port = 5432,
  dbname = "bohemia"
)
}
\arguments{
\item{formatted_data}{A list of dataframes (just one actually) as generated by format_enumerations}

\item{con}{The database connection}

\item{host}{The endpoint}

\item{user}{The database user}

\item{pass}{The database password}

\item{port}{the database bort}

\item{dbname}{The database name}
}
\value{
Data will be updated in the database
}
\description{
Take the formatted enumerations data (as generated by "format_enumerations") and update the data in the database
}
