% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_mopeia_locality_names.R
\name{update_mopeia_locality_names}
\alias{update_mopeia_locality_names}
\title{Update Mopeia locality names}
\usage{
update_mopeia_locality_names(x)
}
\arguments{
\item{x}{The vector to be modified}
}
\value{
A character vector of identical length to \code{x}
}
\description{
Mopeia locality names changed from 2016 to 2017. Specifically, the number of localities were reduced by absorbing some localities (now non-existent) into others. This function is meant to take a vector of "old" localities, and when applicable, update them to the name naming schema
}
