% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_odkx.R
\name{update_odkx}
\alias{update_odkx}
\title{Update ODKX}
\usage{
update_odkx(
  suitcase_file,
  bohemia_dir,
  cloud_endpoint = "https://databrew.app",
  username = "dbrew",
  password = "admin",
  reset_server = TRUE,
  push_new_forms = TRUE
)
}
\arguments{
\item{suitcase_file}{The path to the ODK suitcase jar file}

\item{bohemia_dir}{The path to the cloned, updated bohemia github repo}

\item{cloud_endpoint}{The cloud endpoint}

\item{username}{The username}

\item{password}{The password}

\item{reset_server}{Whether to reset the server}

\item{push_new_forms}{Whether to push new forms to the server}
}
\value{
Data will be updated
}
\description{
Update the ODKX server with new forms
}
