% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_va.R
\name{update_va}
\alias{update_va}
\title{Update va data in the bohemia database}
\usage{
update_va(
  formatted_data,
  con = NULL,
  host = NULL,
  user = NULL,
  pass = NULL,
  port = 5432,
  dbname = "bohemia"
)
}
\arguments{
\item{formatted_data}{A list of dataframes (just one actually) as generated by format_va}

\item{con}{The database connection}

\item{host}{The endpoint}

\item{user}{The database user}

\item{pass}{The database password}

\item{port}{the database bort}

\item{dbname}{The database name}
}
\value{
Data will be updated in the database
}
\description{
Take the formatted va data (as generated by "format_va") and update the data in the database
}
