% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voronoi.R
\name{voronoi}
\alias{voronoi}
\title{Create voronoi tiles}
\usage{
voronoi(shp, poly = NULL)
}
\arguments{
\item{shp}{A spatial points dataframe with the following columns in the data slot: lng, lat, id (the id being the repeated area signifier)}

\item{poly}{An optional spatialPolygonsDataFrame by which the triangles will be trimmed}
}
\value{
A spatialPolygonsDataFrame
}
\description{
Create voronoi tiles using Delaunay triangulation for an entire surface based on ID'ed points
}
