% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_buffers.R
\name{create_buffers}
\alias{create_buffers}
\title{Create buffers}
\usage{
create_buffers(shp, meters = 1000, crs = "+init=epsg:3347",
  ids = NULL)
}
\arguments{
\item{shp}{The spatialPolygonsDataFrame (projected in lat/long)}

\item{meters}{The size of the buffer in meters}

\item{crs}{The coordinate reference system for projecting the data}

\item{ids}{A vector of ids (the same length as the number of rows of shp) with intervention status}
}
\value{
A SpatialPolygonsDataFrame
}
\description{
Generate buffers of size \code{meters} around points or polygons
}
