% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data_dictionary.R
\name{generate_data_dictionary}
\alias{generate_data_dictionary}
\title{Generate data dictionary}
\usage{
generate_data_dictionary(path, language = "English")
}
\arguments{
\item{language}{Language (default English)}
}
\value{
A data dictionary
}
\description{
Generate a data dictionary from a XLSform
}
