% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_generation_va.R
\name{list_generation_va}
\alias{list_generation_va}
\title{VA list generation}
\usage{
list_generation_va(
  death_data,
  keyfile = NULL,
  keyfile_public = NULL,
  location_code = NULL,
  output_file = NULL,
  fake = FALSE
)
}
\arguments{
\item{death_data}{A dataframe in the format of the death data from the censu}

\item{keyfile}{The path to the private key}

\item{keyfile_public}{The path to the public key}

\item{location_code}{A three letter location code (or character vector of multiple codes). If NULL, all locations to be used}

\item{output_file}{The name of the spreadsheet to be written (csv file); if null, will return the table in memory}

\item{fake}{Whether or not to use fake data}
}
\value{
A table will be created
}
\description{
Render a VA list
}
