% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odk_create_location_choices.R
\name{odk_create_location_choices}
\alias{odk_create_location_choices}
\title{Create location choices}
\usage{
odk_create_location_choices(
  country = NULL,
  add_other = TRUE,
  add_ids = FALSE,
  add_codes = TRUE,
  other_word = "Other",
  other_only_levels = c("Village", "Hamlet"),
  lower_it = TRUE
)
}
\arguments{
\item{country}{Which country to use. If \code{NULL} (the default), both will be used.}

\item{add_other}{Whether to add an "other" option}

\item{add_ids}{Whether to add an ID section. Defunct. Keep as FALSE}

\item{add_codes}{Whether to add the hamlet code. TRUE as default}

\item{other_word}{The name of the word to be used to mark "Other"}

\item{other_only_levels}{Whether the other option should be applied only to certain geographical levels. If \code{NULL}, all levels get an "other" option. Otherwise, only the named vector.}

\item{lower_it}{Whether to lowercase the names of the levels (ie "village", not "Village"). Default is False}
}
\value{
A list of two tables named "survey" and "choices"
}
\description{
Generate cascading options for location choices appropriate for the correct location hierarchy in ODK
}
