% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_worker_qrs.R
\name{print_worker_qrs}
\alias{print_worker_qrs}
\title{Print worker QRs}
\usage{
print_worker_qrs(
  wid = seq(0, 999, 1),
  size = 2,
  n = 1,
  output_dir = NULL,
  output_file = "qrs.pdf"
)
}
\arguments{
\item{wid}{The 3 digit id of the worker (or a vector of IDs)}

\item{size}{The size of the image}

\item{n}{The number of images per QR}

\item{output_dir}{The directory to which the file should be written. If
\code{NULL} (the default), the current working directory will be used.}

\item{output_file}{The name of the file to be written.}
}
\value{
A pdf will be written
}
\description{
Render a document of pds for a worker
}
